requirements.txt
pandas
numpy
scipy
scikit-learn

psycopg2-binary
pyyaml

kafka-python

fastapi
uvicorn

streamlit
requests